backend/requirements.txt
fastapi
uvicorn
pillow
torch
opencv-python
transformers
segment-anything
open3d
